0.1.8.0_PR08: Preview 08
Files: Changes to DirectoryTreeEntryNode and DirectoryEntryInfo to support improved handling of cases where the oldest file/directory date time changes and reorders the nodes within a parent directory.  New logic now correctly re-sorts these in parent node.  Added optional pruned path collection list for use in unit test code.  Changed handling of node removal logic when node update finds that file/directory no longer exists.

Logging: Added NullBasicLogger.

All: code changes to address sub-set of lint style compiler comments/messages.

Actions: Changed logic (and signature) for GetFirstFailureOrAllSuccessState EM and added GetFirstFailureOrAllSuccessResultCode EM.  Added VerifySucceeded EM and ActionFailedException (used in unit test).  Added/internally extracted SetStateIssued for use in Action related unit tests.  Changed ActionImplBase to support use with null actionQ (for action related unit tests).  Corrected spelling of ClientFacet ExtensionMethods class.

ValueContainer: Refactored object specific behavior of GetValue variants to support nullable (boxed) value types (structs).  Added additional SetValueL/LS and CreateL/LS variants.

NVS: Changes to ConditionalSetKeyword and ConditionalSetValue to implicitly add the condition that the given name is neither null nor empty to the condition evaluation logic.

Config: Config provider base code now adds ProviderFlags as LS of indicated Provider Flags (IsFixed, ValuesArePersisted, SupportsEE) to provider meta data.

ConfigItemAttribute: Added optional Type, AutoType, Default, AutoDefault, Min, Max UnitOfMeasure, Description, and Custom attributes which may be used to add corresponding config key specific meta data items to related keys.  AutoType and AutoDefault may be used when Type/Default are not specified and will attempt to set the corresponding MD keys based on the initial value and type from the value set itself.

Remoting: Changed "New client session added ..." message by adding partial description of the client's transport endpoint - mainly the IP address and port number from the client side of the session.

E005Data: Added NVS and NV shorthand builders.

E039: Changed E039ObjectID.Empty to specify non-null FullName property.
E040, E094: Added CJState enum and normalized PRJobState enum by adding Removed element.
E090: Added new optional SubstrateJobState.WaitingForPrepare state with corresponding changes to E090SubstrateScheduling source.  Changed common Skipped and Paused detection logic for Stopping and Aborting cases.
E090SubstrateTestingTools: Changes to block triggering prepare requests when PrepareFailedAnnunciator is signaling.  Added AbortJobs, StopJobs and PauseJobs example actions for PrepareFailedAnnunciator.

StringBuilder: Added ConditionalAppend and ConditionalAppendWithDelimiter EMs.

Motion: correction for handling of accel/decel limited case in PointToPointTrajectory.

nuget: updated nuget packages: Newtonsoft.Json to 13.0.1, K4os.Compression.LZ4 to 1.2.16, MessagePack to 2.3.85, SharpZipLib to 1.3.3 and System.Runtime.CompilerServices.Unsafe to 6.0.0

ExtractMDRFtoYYY: readme and usage comment change to explicitly list support for .mdrf2.gz files.

Tools/Compression: Added support for new TreatAllErrorsAsEndOfFile property.

Tools/JsonDotNet:  Added EMs: vc.ConvertToJToken, set.ConvertToJArray, nv.ConvertToJProperty, and nvs.ConvertToJObject to support reverse conversion for use with existing ConvertToVC (et. al.) EMs.

MDRF: Added new TreatAllErrorsAsEndOfFile MDRF2 reading behavior.  Added new MDRF2DecodingIssueHandlingBehavior as SkipToEndOfRecord (default), TreatAsEndOfFile and ThrowDecodingIssueException.  Changed MDRF2QuerySpec StartDateTime and EndDateTime property setters by explicitly converting given DateTime values ToUniversalTime.  Added MDRF2DecodingIssueHandlingBehavior to same and changed default MDRF2FileReaderBehavior to include TreatAllErrorsAsEndOfFile by default.  Changed GetMDRF2FileDigest signature to add new optional keepInlineIndexRecords parameter.  Changed MDRF2FileReadingHelper construction to pass reader behavior values as nullables so that constructor can detect this case and use static (client configurable) fallback values in their place (rather than hard coded defaults).  Corrected issue with MDRF2Writer was not updating its published FileSize.  Added Enable flag to MDRF2LogMessageHandlerAdapter so it can be enabled/disabled when needed.

+++ Base/File/DirectoryTreePruningManager.cs
@@ -21,12 +21,9 @@
  */
 
 using System;
-using System.Collections;
 using System.Collections.Generic;
-using System.IO;
 using System.Linq;
 
-using MosaicLib.Modular.Part;
 using MosaicLib.Utils;
 
 namespace MosaicLib.File
@@ -235,9 +232,12 @@ namespace MosaicLib.File
         /// <summary>This emitter container contains the emitter that is used for all debug messages</summary>
         [Logging.MesgEmitterProperty]
         public Logging.MesgEmitterContainer Debug { get; set; }
+        /// <summary>This emitter container contains the emitter that is used for all trace messages</summary>
+        [Logging.MesgEmitterProperty]
+        public Logging.MesgEmitterContainer Trace { get; set; }
 
         /// <summary>
-        /// This set only property may be used to set the object's emitter containers to contain the desired emitters.  There are three named emitters: Issue, Info, and Debug
+        /// This set only property may be used to set the object's emitter containers to contain the desired emitters.  The following are the named emitters: Issue, Info, Debug, Trace
         /// </summary>
         public IDictionary<string, Logging.IMesgEmitter> EmitterDictionary
         {
@@ -254,6 +254,7 @@ namespace MosaicLib.File
                     Issue = new Logging.MesgEmitterContainer(Logger.Info);
                     Info = new Logging.MesgEmitterContainer(Logger.Info);
                     Debug = new Logging.MesgEmitterContainer(Logger.Debug);
+                    Trace = new Logging.MesgEmitterContainer(Logger.Trace);
                 }
             }
         }
@@ -274,7 +275,7 @@ namespace MosaicLib.File
         private Time.QpcTimer blockPruningUntilAfterTimer = new MosaicLib.Time.QpcTimer() { SelectedBehavior = MosaicLib.Time.QpcTimer.Behavior.NewDefault };
 
         /// <summary>Returns true if pruning is currently blocked for a period of time</summary>
-        protected bool IsPruningBlocked { get { return (blockPruningUntilAfterTimer.Started && !blockPruningUntilAfterTimer.IsTriggered); } }
+        public bool IsPruningBlocked { get { return (blockPruningUntilAfterTimer.Started && !blockPruningUntilAfterTimer.IsTriggered); } }
 
         /// <summary>Sets the timer that is used to block pruning for the given blockPeriod TimeSpan</summary>
         protected void BlockPruningForPeriod(TimeSpan blockPeriod)
@@ -282,7 +283,13 @@ namespace MosaicLib.File
             if (blockPeriod > TimeSpan.Zero)
                 blockPruningUntilAfterTimer.Start(blockPeriod);
             else
-                blockPruningUntilAfterTimer.Stop();
+                ResetPruningBlock();
+        }
+
+        /// <summary>Releases any time based prune holdoff by stopping the blockPruningUntilAfterTimer.</summary>
+        public void ResetPruningBlock()
+        {
+            blockPruningUntilAfterTimer.Stop();
         }
 
         #endregion
@@ -334,7 +341,7 @@ namespace MosaicLib.File
             {
                 string treePruningNeededReason = TreePruningNeededReason;
 
-                if (!treePruningNeededReason.IsNullOrEmpty())
+                if (treePruningNeededReason.IsNeitherNullNorEmpty())
                 {
                     if (PerformIncrementalPrune(treePruningNeededReason))
                         didAnything = true;
@@ -387,12 +394,15 @@ namespace MosaicLib.File
             }
         }
 
-        /// <summary>Gives the total size of this tree and all of the sub-nodes under this node</summary>
-        public Int64 TreeContentsSize { get { Service(false); return treeRootEntry.TreeContentsSize; } }
         /// <summary>Gives the oldest file creation time or directory modification time for this tree or any of its sub-nodes</summary>
         public DateTime OldestUtcTime { get { Service(false); return treeRootEntry.OldestUtcTime; } }
+
+        /// <summary>Gives the total size of this tree and all of the sub-nodes under this node</summary>
+        public Int64 TreeContentsSize { get { Service(false); return treeRootEntry.TreeContentsSize; } }
+
         /// <summary>Gives the total number of items under and including this tree root node.</summary>
         public int TreeItemCount { get { Service(false); return treeRootEntry.TreeItemCount; } }
+        
         /// <summary>Gives the total number of file items under and including this tree.</summary>
         public int TreeFileCount { get { Service(false); return treeRootEntry.TreeFileCount; } }
 
@@ -451,12 +461,15 @@ namespace MosaicLib.File
         /// </summary>
         /// <param name="pruneItemList">List to which new that have been removed from the tree are added - these items must be deleted from the file system separately</param>
         /// <param name="issueEmitter">Gives the IMesgEmitter to which error messages and other Tree traversal issues are given during this process.</param>
-        public void ExtractNextListOfIncrementalItemsToPrune(List<DirectoryEntryInfo> pruneItemList, Logging.IMesgEmitter issueEmitter)
+        /// <param name="traceEmitter">Gives the IMesgEmitter to which some likely repetitious messages will be reported.</param>
+        public void ExtractNextListOfIncrementalItemsToPrune(List<DirectoryEntryInfo> pruneItemList, Logging.IMesgEmitter issueEmitter, Logging.IMesgEmitter traceEmitter = null)
         {
             if (config.PruneMode == PruneMode.PruneDirectories)
-                treeRootEntry.AppendAndRemoveOldestTreeDirectory(pruneItemList, config.MaxEntriesToDeletePerIteration, issueEmitter);
+                treeRootEntry.AppendAndRemoveOldestTreeDirectory(pruneItemList, config.MaxEntriesToDeletePerIteration, issueEmitter, traceEmitter);
             else
-                treeRootEntry.AppendAndRemoveOldestTreeItem(pruneItemList, issueEmitter);
+                treeRootEntry.AppendAndRemoveOldestTreeItem(pruneItemList, issueEmitter, traceEmitter);
+
+            treeRootEntry.UpdateTree(issueEmitter);
         }
 
         /// <summary>
@@ -467,8 +480,12 @@ namespace MosaicLib.File
         /// <param name="deleteEmitter">Gives the IMesgEmitter that will recieve messages about the successful deletions</param>
         /// <param name="issueEmitter">Gives the IMesgEmitter that will receive any messages about failures while attempting to delete each item.</param>
         /// <param name="reason">Gives the string description for the reason that these items are to be deleted.  Null or Empty will be replaced with [NoReasonGiven]</param>
+        /// <param name="traceEmitter">Gives the IMesgEmitter that will be used when reporting the inability to delete a file because it is missing.  When null the <paramref name="issueEmitter"/> will be used instead.</param>
         /// <returns>The number of items that were successfully deleted.</returns>
-        public int DeletePrunedItems(List<DirectoryEntryInfo> pruneItemList, Logging.IMesgEmitter deleteEmitter, Logging.IMesgEmitter issueEmitter, string reason = null)
+        /// <remarks>
+        /// NOTE: the <paramref name="traceEmitter"/> may be used for repeated messages as pruning failures may be reported more than once if the underlying issue is not resolved.
+        /// </remarks>
+        public int DeletePrunedItems(List<DirectoryEntryInfo> pruneItemList, Logging.IMesgEmitter deleteEmitter, Logging.IMesgEmitter issueEmitter, string reason = null, Logging.IMesgEmitter traceEmitter = null)
         {
             int deletedItemCount = 0;		// actually the count of the number of items that we have attempted to delete
 
@@ -484,9 +501,16 @@ namespace MosaicLib.File
                 {
                     try
                     {
-                        System.IO.File.Delete(entryToDelete.Path);
-                        deletedItemCount++;
-                        deleteEmitter.Emit("Pruned file:'{0}', size:{1}, age:{2:f6} days, reason:{3}", entryToDelete.Path, entryToDelete.Length, ageInDays, reason);
+                        if (entryToDelete.IsExistingFile)
+                        {
+                            System.IO.File.Delete(entryToDelete.Path);
+                            deletedItemCount++;
+                            deleteEmitter.Emit("Pruned file:'{0}', size:{1}, age:{2:f6} days, reason:{3}", entryToDelete.Path, entryToDelete.Length, ageInDays, reason);
+                        }
+                        else
+                        {
+                            (traceEmitter ?? issueEmitter).Emit("Unable to prune non-existant file:'{0}', size:{1}, age:{2:f6} days, reason:{3}", entryToDelete.Path, entryToDelete.Length, ageInDays, reason);
+                        }
                     }
                     catch (System.Exception ex)
                     {
@@ -497,9 +521,16 @@ namespace MosaicLib.File
                 {
                     try
                     {
-                        System.IO.Directory.Delete(entryToDelete.Path);
-                        deletedItemCount++;
-                        deleteEmitter.Emit("Pruned directory:'{0}', size:{1}, age:{2:f6} days, reason:{3}", entryToDelete.Path, entryToDelete.Length, ageInDays, reason);
+                        if (entryToDelete.IsExistingDirectory)
+                        {
+                            System.IO.Directory.Delete(entryToDelete.Path);
+                            deletedItemCount++;
+                            deleteEmitter.Emit("Pruned directory:'{0}', size:{1}, age:{2:f6} days, reason:{3}", entryToDelete.Path, entryToDelete.Length, ageInDays, reason);
+                        }
+                        else
+                        {
+                            (traceEmitter ?? issueEmitter).Emit("Unable to prune non-existant directory:'{0}', size:{1}, age:{2:f6} days, reason:{3}", entryToDelete.Path, entryToDelete.Length, ageInDays, reason);
+                        }
                     }
                     catch (System.Exception ex)
                     {
@@ -522,7 +553,7 @@ namespace MosaicLib.File
         /// </summary>
         /// <returns>true if any tree items where deleted, false otherwise.</returns>
         /// <remarks>Number of deletions is limited by config.maxEntriesToDeletePerIteration</remarks>
-        public bool PerformIncrementalPrune(string initialTreePruningNeededReason = null)
+        public bool PerformIncrementalPrune(string initialTreePruningNeededReason = null, List<DirectoryEntryInfo> prunedItemList = null)
         {
             int deletedItemCount = 0;		// actually the count of the number of items that we have attempted to delete
             bool incrementalPruneFailed = false;
@@ -536,10 +567,10 @@ namespace MosaicLib.File
             {
                 pruneItemList.Clear();
 
-                ExtractNextListOfIncrementalItemsToPrune(pruneItemList, Info.Emitter);
+                ExtractNextListOfIncrementalItemsToPrune(pruneItemList, Info.Emitter, Trace.Emitter);
 
                 int iterationItemsToDelete = pruneItemList.Count;
-                int iterationDeletedItemsCount = DeletePrunedItems(pruneItemList, Info.Emitter, Issue.Emitter, reason: treePruningNeededReason);
+                int iterationDeletedItemsCount = DeletePrunedItems(pruneItemList, Info.Emitter, Issue.Emitter, traceEmitter: Trace.Emitter, reason: treePruningNeededReason);
                 deletedItemCount += iterationDeletedItemsCount;
 
                 if (iterationItemsToDelete == 0)
@@ -552,6 +583,9 @@ namespace MosaicLib.File
                     Debug.Emitter.Emit("DeleteItems was not able to delete all of the items during this iteration: items:{0}, deleted items:{1}", iterationItemsToDelete, iterationDeletedItemsCount);
                     incrementalPruneFailed = true;
                 }
+
+                if (prunedItemList != null && pruneItemList.Count > 0)
+                    prunedItemList.AddRange(pruneItemList);
 
                 treePruningNeededReason = TreePruningNeededReason;
             }
@@ -696,4 +730,4 @@ namespace MosaicLib.File
 
     #endregion
 
-} // namespace MosaicLib.File
+}
